--- app/javascript/controllers/index.js
@@ -1,12 +1,0 @@
-import { Application } from "@hotwired/stimulus"
-import { definitionsFromContext } from "@hotwired/stimulus-webpack-helpers"
-
-const application = Application.start()
-const context = require.context(".", true, /\.js$/)
-application.load(definitionsFromContext(context))
-
-// Configure Stimulus development experience
-application.debug = false
-window.Stimulus = application
-
-export { application }
