+++ app/javascript/hello_controller.js
@@ -1,4 +1,3 @@
-// app/javascript/controllers/hello_controller.js
 import { Controller } from "@hotwired/stimulus"
 
 export default class extends Controller {
