+++ app/javascript/packs/application.js
@@ -1,16 +1,19 @@
 // app/javascript/packs/application.js
-import Rails from "@rails/ujs"
-import "@hotwired/turbo-rails"
 import { Application } from "@hotwired/stimulus"
-import { definitionsFromContext } from "@hotwired/stimulus-webpack-helpers";
-Rails.start()
+import { definitionsFromContext } from "@hotwired/stimulus-webpack-helpers"
+import PriceController from "../price_controller"
 
 const application = Application.start()
-const context = require.context("controllers", true, /\.js$/)
-application.load(definitionsFromContext(context))
 
 // Configure Stimulus development experience
 application.debug = false
 window.Stimulus = application
+
+// Import all controllers defined in the importmap under controllers/*
+const context = require.context(".", true, /\.js$/)
+application.load(definitionsFromContext(context))
+
+// Register the price controller
+application.register("price", PriceController)
 
 export { application }
